Map accounts to AccountStore
